requirements_bg_removal.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.21.0
scikit-image>=0.19.0
